+++ dart/lib/src/protocol/sentry_event.dart
@@ -11,6 +11,11 @@ class SentryEvent {
   SentryEvent({
     SentryId eventId,
     DateTime timestamp,
+    Map<String, String> modules,
+    Map<String, String> tags,
+    Map<String, dynamic> extra,
+    List<String> fingerprint,
+    List<Breadcrumb> breadcrumbs,
     this.sdk,
     this.platform,
     this.logger,
@@ -18,7 +23,6 @@ class SentryEvent {
     this.release,
     this.dist,
     this.environment,
-    this.modules,
     this.message,
     this.transaction,
     this.throwable,
@@ -26,17 +30,18 @@ class SentryEvent {
     this.exception,
     this.level,
     this.culprit,
-    this.tags,
-    this.extra,
-    this.fingerprint,
     this.user,
     Contexts contexts,
-    this.breadcrumbs,
     this.request,
     this.debugMeta,
   })  : eventId = eventId ?? SentryId.newId(),
         timestamp = timestamp ?? getUtcDateTime(),
-        contexts = contexts ?? Contexts();
+        contexts = contexts ?? Contexts(),
+        modules = modules != null ? Map.from(modules) : null,
+        tags = tags != null ? Map.from(tags) : null,
+        extra = extra != null ? Map.from(extra) : null,
+        fingerprint = fingerprint != null ? List.from(fingerprint) : null,
+        breadcrumbs = breadcrumbs != null ? List.from(breadcrumbs) : null;
 
   /// Refers to the default fingerprinting algorithm.
   ///
@@ -190,7 +195,7 @@ class SentryEvent {
         release: release ?? this.release,
         dist: dist ?? this.dist,
         environment: environment ?? this.environment,
-        modules: modules ?? this.modules,
+        modules: (modules != null ? Map.from(modules) : null) ?? this.modules,
         message: message ?? this.message,
         transaction: transaction ?? this.transaction,
         throwable: throwable ?? this.throwable,
@@ -198,12 +203,14 @@ class SentryEvent {
         stackTrace: stackTrace ?? this.stackTrace,
         level: level ?? this.level,
         culprit: culprit ?? this.culprit,
-        tags: tags ?? this.tags,
-        extra: extra ?? this.extra,
-        fingerprint: fingerprint ?? this.fingerprint,
+        tags: (tags != null ? Map.from(tags) : null) ?? this.tags,
+        extra: (extra != null ? Map.from(extra) : null) ?? this.extra,
+        fingerprint: (fingerprint != null ? List.from(fingerprint) : null) ??
+            this.fingerprint,
         user: user ?? this.user,
         contexts: contexts ?? this.contexts,
-        breadcrumbs: breadcrumbs ?? this.breadcrumbs,
+        breadcrumbs: (breadcrumbs != null ? List.from(breadcrumbs) : null) ??
+            this.breadcrumbs,
         sdk: sdk ?? this.sdk,
         request: request ?? this.request,
         debugMeta: debugMeta ?? this.debugMeta,
@@ -305,9 +312,8 @@ class SentryEvent {
     }
 
     if (breadcrumbs != null && breadcrumbs.isNotEmpty) {
-      json['breadcrumbs'] = <String, List<Map<String, dynamic>>>{
-        'values': breadcrumbs.map((b) => b.toJson()).toList(growable: false)
-      };
+      json['breadcrumbs'] =
+          breadcrumbs.map((b) => b.toJson()).toList(growable: false);
     }
 
     json['sdk'] = sdk?.toJson() ??