+++ flutter/lib/sentry_flutter.dart
@@ -2,4 +2,5 @@
 export 'package:sentry/sentry.dart';
 
 export 'src/default_integrations.dart';
+export 'src/navigation/sentry_navigation_observer.dart';
 export 'src/sentry_flutter.dart';