+++ flutter/lib/src/platform_checker.dart
@@ -1,0 +1,5 @@
+import 'dart:io';
+
+/// verify if the platform is iOS
+/// used to run loadContextsIntegration only on iOS
+bool isIOS() => Platform.isIOS;