+++ flutter/test/load_contexts_integrations_test.dart
@@ -5,11 +5,11 @@ import 'package:sentry_flutter/sentry_flutter.dart';
 import 'mocks.dart';
 
 void main() {
-  const MethodChannel _channel = MethodChannel('sentry_flutter');
+  const _channel = MethodChannel('sentry_flutter');
 
   TestWidgetsFlutterBinding.ensureInitialized();
 
-  bool called = false;
+  var called = false;
 
   setUp(() {
     _channel.setMockMethodCallHandler((MethodCall methodCall) async {
@@ -106,8 +106,8 @@ void main() {
       final eventSdk = SdkVersion(
         name: 'sdk1',
         version: '1.0',
-        integrations: ['EventIntegration'],
-        packages: [const SentryPackage('event-package', '2.0')],
+        integrations: const ['EventIntegration'],
+        packages: const [SentryPackage('event-package', '2.0')],
       );
       final e = SentryEvent(sdk: eventSdk);
       final event = await options.eventProcessors.first(e, null);